requirements.txt
fastapi
uvicorn
python-dotenv
mysql
PyMySQL
aiomysql
# mysqlclient
sqlalchemy
cryptography
celery
redis
pandas
openpyxl
python-multipart
matplotlib
python-multipart
requests
wkhtmltopdf